Optimize block storage speed
Mojang's new block storage format was really slow, and hurt performance.
Even with an optimized build it took 20%+ of the runtime.
Switching to the legacy system (storing blocks as u16) gave a large speedup.
This comes at a slight memory cost, but even with 1000 chunks loaded
it'll only take 8MB.